backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
tensorflow>=2.16,<3.0
numpy>=1.26,<3.0
Pillow>=10.3,<11.0
python-multipart>=0.0.9,<1.0
